bug fix: transitions with diff RT
If the q values are the same for 2 different runs than behaviour is
undefined for which precursor selecting. This can mean that transitions
part of the same transition group have different RT/IM. To address this,
also sort by RunId. If Q values are the same just take the first run